Split demo into people and dino demos (#118)
* split demos/demo into dino and people demos

* pr refinements

* fix router listen log from error to info

* changes based on PR review

* changes based on PR review
