+++ demos/dino/src/main/java/com/nordstrom/xrpc/demos/dino/DinoGetRequestEncoder.java
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package com.nordstrom.xrpc.demo;
+package com.nordstrom.xrpc.demos.dino;
 
-import com.nordstrom.xrpc.demo.proto.Dino;
+import com.nordstrom.xrpc.demos.dino.proto.DinoGetRequest;
+
 import java.io.IOException;
 
-public class DinoDecoder {
+public class DinoGetRequestEncoder {
   public static void main(String[] args) throws IOException {
-    byte[] bytes = new byte[System.in.available()];
-    System.in.read(bytes, 0, bytes.length);
-    Dino dino = Dino.parseFrom(bytes);
-    System.out.println(dino);
+    DinoGetRequest dinoGetRequest = DinoGetRequest.newBuilder().setName(args[0]).build();
+    System.out.write(dinoGetRequest.toByteArray());
   }
 }
