+++ demos/dino/src/main/resources/demo.conf
