+++ demos/dino/src/main/resources/logback.groovy
