+++ demos/people/build.gradle
@@ -5,20 +5,27 @@ buildscript {
     jcenter()
   }
   dependencies {
-    classpath 'com.google.protobuf:protobuf-gradle-plugin:0.8.3'
+    classpath 'org.junit.platform:junit-platform-gradle-plugin:1.0.1'
     classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.1'
   }
 }
 
+plugins {
+  id 'com.github.sherter.google-java-format' version '0.6'
+}
+
 apply plugin: 'java'
-apply plugin: 'com.google.protobuf'
 apply plugin: 'com.github.johnrengelman.shadow'
+
+// Junit 5 gradle runner.
+apply plugin: 'org.junit.platform.gradle.plugin'
+
 
 group = 'com.nordstrom.xrpc'
 version = '0.1.1-SNAPSHOT'
-archivesBaseName = 'xrpc-demo'
+archivesBaseName = 'xrpc-people-demo'
 
-description = """xrpc demo project."""
+description = """xrpc people demo project."""
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
@@ -39,7 +46,9 @@ dependencies {
   compile 'io.netty:netty-all:4.1.19.Final'
   compile 'org.projectlombok:lombok:1.16.16'
   compile 'org.slf4j:slf4j-api:1.7.25'
-  testCompile 'junit:junit:4.12'
+  testCompile 'org.junit.jupiter:junit-jupiter-api:5.0.1'
+  testCompile 'org.apiguardian:apiguardian-api:1.0.0'
+  testRuntime 'org.junit.jupiter:junit-jupiter-engine:5.0.1'
   runtimeOnly 'ch.qos.logback:logback-classic:1.1.7'
   runtimeOnly 'ch.qos.logback:logback-core:1.1.7'
   runtimeOnly 'org.codehaus.groovy:groovy-all:2.4.1'
@@ -53,16 +62,8 @@ sourceSets {
   }
 }
 
-// Configure protobuf build.
-protobuf {
-  // Configure the protoc executable to use the official distribution.
-  protoc {
-    artifact = 'com.google.protobuf:protoc:3.0.0'
-  }
-}
-
 shadowJar {
   manifest {
-    attributes 'Main-Class': 'com.nordstrom.xrpc.demo.Example'
+    attributes 'Main-Class': 'com.nordstrom.xrpc.demos.people.Application'
   }
 }
