+++ demos/people/src/main/resources/demo.conf
@@ -1,0 +1,23 @@
+# Place any application-level configuration here.
+mykey = "value"
+
+# Place any xrpc configuration overrides here.
+xrpc {
+  worker_name_format = "demo-server-worker-%d"
+  soft_req_per_sec = 500.0
+  hard_req_per_sec = 550.0
+  console_reporter = true
+
+  # IP Black List
+  enable_black_list = false
+  ip_black_list = []
+
+  # IP White List
+  enable_white_list = true
+  ip_white_list = ["127.0.0.1"]
+
+  cors {
+    enable = true
+    allowed_origins = ["test.domain"]
+  }
+}
