+++ demos/people/src/main/resources/logback.groovy
@@ -1,0 +1,16 @@
+import ch.qos.logback.classic.filter.ThresholdFilter
+
+appender("CONSOLE", ConsoleAppender) {
+    withJansi = true
+
+    filter(ThresholdFilter) {
+        level = DEBUG
+    }
+
+    encoder(PatternLayoutEncoder) {
+        pattern = "%-4relative [%thread] %-5level %logger{30} - %msg%n"
+        outputPatternAsHeader = false
+    }
+}
+
+root(INFO, ["CONSOLE"])
