+++ settings.gradle
@@ -1,5 +1,8 @@
 include 'xrpc'
-include 'demos:demo'
+include 'xrpc-testing'
+include 'demos:dino'
+include 'demos:people'
 include 'demos:request-response-pipeline'
 
 rootProject.name = 'xrpc'
+
