+++ xrpc-testing/build.gradle
@@ -10,17 +10,7 @@ buildscript {
 
 plugins {
   id 'com.github.sherter.google-java-format' version '0.6'
-  id 'net.researchgate.release' version '2.4.0'
   id 'com.github.spotbugs' version '1.6.0'
-}
-
-release {
-  preTagCommitMessage = 'Updating to release version '
-  tagCommitMessage = 'New release: '
-  newVersionCommitMessage = 'Creating new SNAPSHOT version '
-  git {
-    pushToRemote = 'upstream'
-  }
 }
 
 apply plugin: 'java'
@@ -28,27 +18,24 @@ apply plugin: 'maven'
 apply plugin: 'maven-publish'
 apply plugin: 'checkstyle'
 
-// Junit 5 gradle runner.
-apply plugin: 'org.junit.platform.gradle.plugin'
-
 group = 'com.nordstrom.xrpc'
 archivesBaseName = 'xrpc'
 
-description = """Simple, production ready Java API server built on top of functional composition."""
+description = """Testing Utilities for simple, production ready Java API server built on top of functional composition."""
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 tasks.withType(JavaCompile) {
-    options.encoding = 'UTF-8'
-    // The tls package uses a ton of proprietary APIs that generate warnings. Suppress these (and fork
-    // the compiler so that the definition works).
-    options.compilerArgs += [
-        '-XDignore.symbol.file',
-        '-Xlint:auxiliaryclass,cast,classfile,deprecation,dep-ann,divzero,empty,fallthrough,' +
-            'finally,options,overloads,overrides,path,rawtypes,static,try,unchecked,varargs',
-        '-Werror']
-    options.fork = true
-    options.forkOptions.executable = 'javac'
+  options.encoding = 'UTF-8'
+  // The tls package uses a ton of proprietary APIs that generate warnings. Suppress these (and fork
+  // the compiler so that the definition works).
+  options.compilerArgs += [
+    '-XDignore.symbol.file',
+    '-Xlint:auxiliaryclass,cast,classfile,deprecation,dep-ann,divzero,empty,fallthrough,' +
+      'finally,options,overloads,overrides,path,rawtypes,static,try,unchecked,varargs',
+    '-Werror']
+  options.fork = true
+  options.forkOptions.executable = 'javac'
 }
 
 repositories {
@@ -57,31 +44,9 @@ repositories {
 }
 
 dependencies {
-  compile 'com.google.auto.value:auto-value:1.5.3'
-  compile 'com.google.guava:guava:21.0'
   compile 'com.squareup.okhttp3:okhttp:3.8.0'
   compile 'com.squareup.okio:okio:1.13.0'
   compile 'com.typesafe:config:1.3.1'
-  compile 'com.xjeffrose:xio:0.13.0'
-  compile 'io.dropwizard.metrics:metrics-core:3.2.5'
-  compile 'io.dropwizard.metrics:metrics-healthchecks:3.2.5'
-  compile 'io.dropwizard.metrics:metrics-json:3.2.5'
-  compile 'com.fasterxml.jackson.core:jackson-databind:2.9.2'
-  compile 'io.netty:netty-all:4.1.19.Final'
-  compile 'io.netty:netty-tcnative-boringssl-static:2.0.7.Final'
-  compile 'io.netty:netty-transport-native-epoll:4.1.19.Final'
-  compile 'io.netty:netty-transport-native-kqueue:4.1.19.Final'
-  compile 'org.projectlombok:lombok:1.16.16'
-  compile 'org.slf4j:slf4j-api:1.7.25'
-  testCompile project(':xrpc-testing')
-  testCompile 'org.junit.jupiter:junit-jupiter-api:5.0.1'
-  testCompile 'org.apiguardian:apiguardian-api:1.0.0'
-  testCompile 'org.mockito:mockito-core:2.13.0'
-  testCompile 'com.squareup.okhttp3:mockwebserver:3.8.1'
-  testRuntime 'org.junit.jupiter:junit-jupiter-engine:5.0.1'
-  runtimeOnly 'ch.qos.logback:logback-classic:1.1.7'
-  runtimeOnly 'ch.qos.logback:logback-core:1.1.7'
-  runtimeOnly 'org.codehaus.groovy:groovy-all:2.4.1'
 }
 
 // Run with the latest checkstyle version. Required for our checkstyle.xml.
@@ -157,7 +122,7 @@ publishing {
 
       pom.withXml {
         asNode().appendNode('description', rootProject.description)
-        asNode().appendNode('name', 'xrpc library')
+        asNode().appendNode('name', 'xrpc testing library')
         asNode().appendNode('url', 'https://github.com/Nordstrom/xrpc')
         asNode().children().last() + pomExtra
       }
