+++ xrpc/src/main/java/com/nordstrom/xrpc/server/XrpcRequest.java
@@ -39,8 +39,10 @@ import java.nio.charset.Charset;
 import java.util.Collections;
 import java.util.Map;
 import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
 
 /** Xprc specific Request object. */
+@Slf4j
 public class XrpcRequest {
   private final ObjectMapper mapper;
 
@@ -205,6 +207,7 @@ public class XrpcRequest {
   }
 
   public FullHttpResponse okJsonResponse(Object body) throws IOException {
+
     return jsonResponse(HttpResponseStatus.OK, body);
   }
